percent as float

+++ src/hp_BH1750.cpp
@@ -553,7 +553,7 @@ bool hp_BH1750::saturated() const
 // switch between this values when requiered
 // If you use the fast, but low qualtiy BH1750_QUALITY_LOW = 0x23, the function will not change this quality.
 
-bool hp_BH1750::adjustSettings(byte percent, bool forcePreShot)
+bool hp_BH1750::adjustSettings(float percent, bool forcePreShot)
 {
   if (_value == BH1750_SATURATED || forcePreShot == true) // If last result is saturated, perfom a measurement at low sensitivity
   {
@@ -572,14 +572,14 @@ bool hp_BH1750::adjustSettings(byte percent, bool forcePreShot)
 // You have to declare the variables BH1750Quality and mtreg bevore calling this function
 // After calling this function this variables contain the appropiate values
 
-void hp_BH1750::calcSettings(unsigned int value, BH1750Quality &qual, byte &mtreg, byte percent)
+void hp_BH1750::calcSettings(unsigned int value, BH1750Quality &qual, byte &mtreg, float percent)
 {
-  if (percent > 100)
-    percent = 100;
+  if (percent > 100.0)
+    percent = 100.0;
   _percent = percent;
   if (value==0)
     value=1;
-  float highBound = value / (float)percent * 100;
+  float highBound = value / percent * 100.0;
   unsigned long newMtreg = (float)BH1750_SATURATED / highBound * mtreg + .5;
   switch (qual)
   {
@@ -604,6 +604,6 @@ void hp_BH1750::calcSettings(unsigned int value, BH1750Quality &qual, byte &mtre
   mtreg = newMtreg;
 }
 
-byte hp_BH1750::getPercent() const {
+float hp_BH1750::getPercent() const {
   return _percent;
 }
